% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{admin_match}
\alias{admin_match}
\title{match admin region}
\usage{
admin_match(admin_unit = NULL, country = NULL, admin_units_seasonal)
}
\arguments{
\item{admin_unit}{Character for admin region. Some fuzzy logic will be used to
match. If not provided then no seasonality is introduced. Default = NULL}

\item{country}{Character for country within which admin unit is in.
Default = NULL}

\item{admin_units_seasonal}{Dataframe of seasonality data for country and admin unit}
}
\description{
\code{admin_match} Matches the user input admin unit and country with data
}
