% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_ivm_functions.R
\name{effective_mortality}
\alias{effective_mortality}
\title{effective_mortality}
\usage{
effective_mortality(op)
}
\arguments{
\item{op}{}
}
\value{

}
\description{
effective_mortality
}
\examples{

}
