% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium-init-multi.R
\name{equilibrium_init_multi}
\alias{equilibrium_init_multi}
\title{Equilibrium initialisation for multiple populations}
\usage{
equilibrium_init_multi(
  age_vector,
  het_brackets,
  country = NULL,
  admin_unit = NULL,
  ft,
  EIR_vector,
  model_param_list
)
}
\arguments{
\item{age_vector}{Vector of age brackets.}

\item{het_brackets}{Integer number of biting heteogenity compartments.}

\item{country}{String for country of interest. If NULL the seasonal parameters
will attempt to be loaded using just the admin unit, however if there is ambiguity
in the admin unit an error will be thrown. If both NULL then no seasonality is
assumed. Default = NULL.}

\item{admin_unit}{String for admin unit with country for loading seasonal
parameters. If country is NULL, the admin unit will attempt to be located,however
if there is ambiguity in the admin unit an error will be thrown. If both country
and admin_unit are NULL then no seasonality is assumed. Default = NULL.}

\item{ft}{Numeric for the frequency of people seeking treatment.}

\item{EIR_vector}{Vector of numeric for desired annual EIR per patch.}

\item{model_param_list}{List of epidemiological parameters created by}
}
\description{
\code{equilibrium_init_multi} creates an equilibrium initialisation state to be
used within later model runs
}
