% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fourier}
\alias{fourier}
\title{Generate Fourier}
\usage{
fourier(x, ss)
}
\arguments{
\item{x}{Vector of times}

\item{ss}{Vector of fourier parameters in this order a0, a1, b1, a2, b2, a3, b3}
}
\description{
\code{fourier} computes fourier series to be used later for seasonality in carrying capacity
}
