% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator_utils.R
\name{grab_user_params}
\alias{grab_user_params}
\title{Grab parameters from odin model}
\usage{
grab_user_params(model_path)
}
\arguments{
\item{model_path}{Character path to odin model}
}
\value{
Vector of characters for required variables
}
\description{
\code{grab_user_params} returns vector of all user required variables from an odin
model path
}
\examples{
head(hanojoel:::grab_user_params(system.file("extdata/odin_model.R",package="hanojoel")))

}
