% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanojoel.R
\docType{package}
\name{hanojoel-package}
\alias{hanojoel}
\alias{hanojoel-package}
\title{hanojoel: Deterministic malaria transmission model}
\description{
hanojoel implements the Imperial College malaria transmission model
using the R package odin. A related suite of tools for setting up
different regional settings and model types is provided provides a flexible grammar of data manipulation.
}
