% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_ivm_functions.R
\name{ivm_fun}
\alias{ivm_fun}
\title{ivm_fun}
\usage{
ivm_fun(
  IVM_start_times,
  time_period,
  hazard_profile,
  ivm_coverage = 0.8,
  ivm_min_age = 5,
  ivm_max_age = 200,
  bites_per_3_days = 1
)
}
\arguments{
\item{bites_per_3_days}{}
}
\value{

}
\description{
ivm_fun
}
\examples{

}
