% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{load_file}
\alias{load_file}
\title{load_file}
\usage{
load_file(name)
}
\arguments{
\item{name}{the name of a file within the inst/extdata folder.}

\item{age}{Vector of age brackets.}
}
\description{
Load a file from within the inst/extdata folder of the
  hanojoel package. File extension must be one of .csv, .txt, or .rds.
}
\details{
\code{load_file} load package file
}
