% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{match_clean}
\alias{match_clean}
\title{match clean}
\usage{
match_clean(a, b)
}
\arguments{
\item{a}{First string to compare. Default = NULL}

\item{b}{Second string to compare. Default = NULL}
}
\description{
match clean
}
