% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_ivm_functions.R
\name{mda_fun}
\alias{mda_fun}
\title{mda_fun}
\usage{
mda_fun(
  ncc = 2,
  MDA_times = NULL,
  MDA_grp_prop = 0.9,
  MDA_cov = 0.8,
  MDA_st_age = 0.5,
  MDA_en_age = 80,
  MDA_succ = 0.95,
  MDA_drug_choice = 2,
  init_age = NULL
)
}
\arguments{
\item{MDA_drug_choice}{}
}
\value{

}
\description{
mda_fun
}
\examples{

}
