% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{Run_Model}
\alias{Run_Model}
\title{Run_Model}
\usage{
Run_Model(age, EIR, ft, admin2, time)
}
\arguments{
\item{age}{Vector of age brackets.}

\item{EIR}{Numeric of annual EIR}

\item{ft}{Numeric of proportion of symptomatic cases recieving treatment}

\item{admin2}{Character of admin unit}

\item{time}{Numeric of length of time that the model will run for in days}
}
\description{
\code{Run_Model} runs model using declared age, EIR, ft, country, admin
}
