% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seasonality}
\alias{seasonality}
\title{Seasonality}
\usage{
seasonality(ss)
}
\arguments{
\item{ss}{Vector of fourier parameters in this order a0, a1, b1, a2, b2, a3, b3}
}
\description{
\code{seasonality} Generate seasonality scaling vector for one year, to be applied to carrying capacity.
}
