% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.R
\name{vivax_model_param_list_create}
\alias{vivax_model_param_list_create}
\title{Model Parameter List Creation: P. vivax}
\usage{
vivax_model_param_list_create(
  eta = 1/(21 * 365),
  rho = 0.85,
  a0 = 2920,
  sigma2 = 1.292285,
  max_age = 100 * 365,
  rLM = 1/10,
  rT = 1,
  rD = 1/5,
  rP = 1/28,
  dPCR_min = 10,
  dPCR_max = 70,
  dE = 10,
  cD = 0.8,
  cT = 0.4,
  cPCR = 0.035,
  cLM = 0.1,
  b = 0.5,
  d_par = 10 * 365,
  u_par = 19.77,
  phi_LM_max = 0.8918,
  phi_LM_min = 0.0043,
  A_LM_50pc = 27.52,
  K_LM = 2.403,
  A_PCR_50pc = 9.9,
  K_PCR = 4.602,
  d_clin = 30 * 365,
  u_clin = 7.85,
  phi_D_max = 0.8605,
  phi_D_min = 0.018,
  A_D_50pc = 11.538,
  K_D = 2.25,
  p_mat = 0.421,
  d_mat = 35.148,
  gamma_L = 1/383,
  ff = 0.02439024,
  delayMos = 8.4,
  tau1 = 0.69,
  tau2 = 2.31,
  mu0 = 1/6,
  Q0 = 0.5,
  chi = 0.9,
  bites_Bed = 0.9,
  bites_Indoors = 0.9,
  muEL = 0.0338,
  muLL = 0.0348,
  muPL = 0.249,
  dEL = 6.64,
  dLL = 3.72,
  dPL = 0.643,
  gammaL = 13.25,
  betaL = 21.19,
  num_int = 1,
  itn_cov = 0,
  irs_cov = 0,
  ITN_IRS_on = -1,
  DY = 365,
  d_ITN0 = 0.3,
  r_ITN0 = 0.6,
  r_ITN1 = 0.2,
  r_IRS0 = 0.4,
  d_IRS0 = 0.5,
  irs_half_life = 0.5 * DY,
  itn_half_life = 2.64 * DY,
  IRS_interval = 1 * DY,
  ITN_interval = 3 * DY,
  stop_rebound_switch = 0,
  hypnozoite_prev_threshold = 1e-05,
  ...
)
}
\arguments{
\item{rho}{Age-dependent biting parameter. Default = 0.85}

\item{a0}{Age-dependent biting parameter. Default = 2920}

\item{sigma2}{Variance of the log heterogeneity in biting rates. Default = 1.29}

\item{max_age}{Maximum age in days. Default = 100*365}

\item{rLM}{Rate of leaving light microscopy-detectable infection. Default = 0.1}

\item{rT}{Rate of leaving treatment. Default = 1}

\item{rD}{Rate of leaving clinical disease. Default = 0.2}

\item{rP}{Rate of leaving prophylaxis. Default = 0.036}

\item{dPCR_min}{Minimum duration of PCR-detectable infection. Default = 10}

\item{dPCR_max}{Maximum duration of PCR-detectable infection. Default = 70}

\item{dE}{Latent period of human infection. Default = 10}

\item{cD}{Untreated disease contribution to human-to-mosquito infectiousness. Default = 0.8}

\item{cT}{Treated disease contribution to human-to-mosquito infectiousness. Default = 0.4}

\item{cPCR}{PCR-detectable infection contribution to infectiousness. Default = 0.035}

\item{cLM}{LM-detectable infection contribution to infectiousness. Default = 0.1}

\item{b}{Mosquito-to-human infection probability. Default = 0.5}

\item{d_par}{Duration of anti-parasite immunity. Default = 3650}

\item{u_par}{Anti-parasite immune boosting refractory period. Default = 19.77}

\item{phi_LM_max}{Probability of LM-detectable infection with no immunity. Default = 0.8918}

\item{phi_LM_min}{Probability of LM-detectable infection with full immunity. Default = 0.0043}

\item{A_LM_50pc}{Anti-parasite immunity for 50 percent reduction in LM-detectable infection. Default = 27.52}

\item{K_LM}{Shape parameter for LM-detectable infection. Default = 2.403}

\item{A_PCR_50pc}{Anti-parasite immunity for 50 percent reduction in duration of PCR-detectable infection. Default = 9.9}

\item{K_PCR}{Shape parameter for duration of PCR-detectable infection. Default = 4.602}

\item{d_clin}{Duration of clinical immunity. Default = 10950}

\item{u_clin}{Clinical immune boosting refractory period. Default = 7.85}

\item{phi_D_max}{Probability of clinical episode with no immunity. Default = 0.8605}

\item{phi_D_min}{Probability of clinical episode with full immunity. Default = 0.018}

\item{A_D_50pc}{Anti-parasite immunity for 50 percent reduction in clinical episode. Default = 11.538}

\item{K_D}{Shape parameter for clinical disease. Default = 2.250}

\item{p_mat}{New-born immunity relative to mother’s. Default = 0.421}

\item{d_mat}{Duration of maternal immunity. Default = 35.148}

\item{gamma_L}{Rate of hypnozoite clearance. Default = 0.0026}

\item{ff}{Rate of relapse. Default = 0.024}

\item{delayMos}{Extrinsic incubation period/duration of sporogony. Default = 8.4}

\item{tau1}{Duration of host seeking, assumed to be constant between species. Default = 0.69}

\item{tau2}{Duration of mosquito resting after feed. Default = 2.31}

\item{mu0}{Daily mortality of adult mosquitoes. Default = 0.167}

\item{Q0}{Anthrophagy probability. Default = 0.5}

\item{chi}{Endophily probability. Default = 0.9}

\item{bites_Bed}{Percentage of bites indoors and in bed. Default = 0.9}

\item{bites_Indoors}{Percentage of bites indoors . Default = 0.9}

\item{muEL}{Per capita daily mortality rate of early stage larvae (low density). Default = 0.0338}

\item{muLL}{Per capita daily mortality rate of late stage larvae (low density). Default = 0.0348}

\item{muPL}{Per capita daily mortality rate of pupae. Default = 0.249}

\item{dEL}{Development time of early stage larvae. Default = 6.64}

\item{dLL}{Development time of late stage larvae. Default = 3.72}

\item{dPL}{Development time of pupae. Default = 0.643}

\item{gammaL}{Relative effect of density dependence on late instars relative to early instars. Default = 13.25}

\item{betaL}{Number of eggs laid per day per mosquito. Default = 21.2}

\item{num_int}{Number of intervention parameters.  Default = 4}

\item{itn_cov}{The proportion of people that use an ITN. Default = 0}

\item{irs_cov}{The proportion of people living in houses that have been sprayed. Default = 0}

\item{ITN_IRS_on}{Time of ITN and IRS to be activated. Default = -1, i.e. never.}

\item{DY}{Duration of year (days). Default = 365}

\item{d_ITN0}{Probability of dying with an encounter with ITN (max). Default = 0.41}

\item{r_ITN0}{Probability of repeating behaviour with ITN (max). Default = 0.56}

\item{r_ITN1}{Probability of repeating behaviour with ITN (min). Default = 0.24}

\item{r_IRS0}{Probability of repeating behaviour with IRS (min). Default = 0.6}

\item{d_IRS0}{Probability of dying with an encounter with IRS (max). Default = 1}

\item{irs_half_life}{IRS half life. Default =   0.5 * DY}

\item{itn_half_life}{ITN half life. Default =   2.64 * DY}

\item{IRS_interval}{How long before IRS is repeated, i.e. when IRS decay = 1. Default =   1 * DY}

\item{ITN_interval}{How long before ITN is repeated, i.e. when IRS decay = 1.  Default =   3 * DY}

\item{stop_rebound_switch}{Set to 1 to prevent a rebound in cases after decades due to relapse from very small hypnozoite reservoir in the population. Default = 0}

\item{hypnozoite_prev_threshold}{Hypnozoite prevalence in the population at which stop_rebound_switch is triggered. Default = 1e-05}

\item{...}{Any other parameters needed for non-standard model. If they share the same name
as any of the defined parameters \code{vivax_model_param_list_create} will stop. You can either write
any extra parameters you like individually, e.g. vivax_model_param_list_create(extra1 = 1, extra2 = 2)
and these parameters will appear appended to the returned list, or you can pass explicitly
the ellipsis argument as a list created before, e.g. vivax_model_param_list_create(...=list(extra1 = 1, extra2 = 2))}
}
\description{
The function \code{vixax_model_param_list_create} creates list of model parameters for
P. vivax to be used within \code{vivax_equilibrium_init_create}
}
