% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanojoel.R
\docType{package}
\name{hanojoel-package}
\alias{hanojoel}
\alias{hanojoel-package}
\title{hanojoel: Deterministic malaria transmission model}
\description{
hanojoel implements the Imperial College malaria transmission model
using the R package odin. A related suite of tools for setting up
different regional settings and model types is provided provides a flexible grammar of data manipulation.
}
\author{
\strong{Maintainer}: OJ Watson \email{o.watson15@imperial.ac.uk}

Authors:
\itemize{
  \item Joel Hellewell \email{j.hellewell14@imperial.ac.uk}
  \item Hannah Slater \email{hannah.slater08@imperial.ac.uk}
  \item Juliette Unwin \email{h.unwin@imperial.ac.uk}
}

}
